+++ CHANGELOG.md
@@ -5,6 +5,74 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 27-Nov-2021
+
+### Added
+
+- Upgrade Egui to 0.15.0 ([#45](https://github.com/mvlabat/bevy_egui/pull/45)).
+
+## [0.7.1] - 06-Oct-2021
+
+### Added
+
+- Add `EguiStartupSystem` system labels.
+
+### Fixed
+
+- Initialize egui contexts during startup (fixes [#41](https://github.com/mvlabat/bevy_egui/issues/41)).
+
+## [0.7.0] - 05-Sep-2021
+
+### Added
+
+- Upgrade Egui to 0.14.0 ([#38](https://github.com/mvlabat/bevy_egui/pull/38)).
+
+## [0.6.2] - 15-Aug-2021
+
+### Fixed
+
+- Fix receiving input when holding a button ([#37](https://github.com/mvlabat/bevy_egui/pull/37)).
+
+## [0.6.1] - 20-Jul-2021
+
+### Fixed
+
+- Fix more edge-cases related to invalid scissors.
+
+## [0.6.0] - 29-Jun-2021
+
+### Added
+
+- Upgrade Egui to 0.13.0.
+
+## [0.5.0] - 22-May-2021
+
+### Added
+
+- Upgrade Egui to 0.12.0.
+
+## [0.4.2] - 03-May-2021
+
+### Added
+
+- Better error message for a missing Egui context ([#24](https://github.com/mvlabat/bevy_egui/pull/24) by @jakobhellermann)
+- Add `try_ctx_for_window` function ([#20](https://github.com/mvlabat/bevy_egui/pull/20) by @jakobhellermann)
+
+## [0.4.1] - 24-Apr-2021
+
+### Fixed
+
+- Fix crashes related to invalid scissor or window size ([#18](https://github.com/mvlabat/bevy_egui/pull/18))
+
+## [0.4.0] - 10-Apr-2021
+
+Huge thanks to @jakobhellermann and @Weasy666 for contributing to this release!
+
+### Added
+
+- Implement Egui 0.11.0 support ([#12](https://github.com/mvlabat/bevy_egui/pull/12) by @Weasy666 and @jakobhellermann).
+- Implement multiple windows support ([#14](https://github.com/mvlabat/bevy_egui/pull/14) by @jakobhellermann).
+
 ## [0.3.0] - 02-Mar-2021
 
 ### Added